提交

+++ 2.Firmware/Mcu1/.pio/libdeps/pico32/Simple FOC/src/sensors/MagneticSensorSPI.h
@@ -3,7 +3,7 @@
 
 
 #include "Arduino.h"
-#include <C:/Users/Administrator/.platformio/packages/framework-arduinoespressif32@src-f10a7ed2306067c023b128ae96ba301b/libraries/SPI/src/SPI.h>
+#include <SPI.h>
 #include "../common/base_classes/Sensor.h"
 #include "../common/foc_utils.h"
 #include "../common/time_utils.h"
@@ -30,7 +30,7 @@ class MagneticSensorSPI: public Sensor{
      * @param bit_resolution   sensor resolution bit number
      * @param angle_register  (optional) angle read register - default 0x3FFF
      */
-    MagneticSensorSPI(int cs, float bit_resolution, int angle_register = 0);
+    MagneticSensorSPI(int cs, int bit_resolution, int angle_register = 0);
     /**
      *  MagneticSensorSPI class constructor
      * @param config   SPI config
